moves tests all back into one file again
